refactor(panels): rename lazy-load targets from Component to Panel files
Align panel chunk naming with ThreeD/Pose so dist-lib outputs identifiable
{PanelName}Panel-*.js chunks instead of generic Component-*.js files.
