+++ src/features/panels/Image/ImagePanel.tsx
