+++ src/features/panels/JointStatePlot/JointStatePlotPanel.tsx
